Break script.js into several files for easier maintainablity
