+++ config/page.json
@@ -1,6 +1,6 @@
 {
   "logo": "assets/logo.png",
-  "backgroundImage": "https://cataas.com/cat?filter=sepia&width=1280&height=720",
+  "backgroundImage": "https://cataas.com/cat?width=1280&height=720",
   "favicon": "https://cdn3.iconfinder.com/data/icons/font-awesome-regular-1/512/face-grin-squint-512.png",
   "title": "Cat - OpenLinkway",
   "description": "Cat's Link site",
