+++ index.html
@@ -7,7 +7,7 @@
     <link rel="shortcut icon" id="favicon" href="assets/favicon.webp" />
     <title>OpenLinkway</title>
     <link rel="stylesheet" href="styles.css" />
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
   </head>
   <body>
     <div id="bg-container"></div>
